+++ packages/DataElement/src/DataElement.ts
@@ -3,13 +3,10 @@ import { event } from "@domx/eventmap/decorators";
 import { Middleware } from "@domx/middleware";
 export {
     customDataElements,
-    customDataElement,
     DataElement,
     DataElementCtor,
     DataElementMetaData,
-    DataProperties,
-    dataProperty,
-    event
+    DataProperties
 };
 import { RootState } from "./RootState";
 
@@ -155,51 +152,6 @@ const customDataElements = {
         window.customElements.define(elementName, element);
     }
 };
-
-
-interface CustomDataElementOptions {
-    /** Sets which property is to be used as the stateId; default: stateId */
-    stateIdProperty?: string,
-    /** Sets the default event listener element for events; default: "self" */
-    eventsListenAt?: EventMapListenAt|string
-}
-/**
- * A class decorator that defines the custom element with
- * `window.customElements.define` and tags the element name
- * for use in RootState.
- * 
- * Options allow for setting `stateIdProperty` and `eventsListenAt`.
- * @param elementName {string}
- * @param options {CustomDataElementOptions}
- */
-const customDataElement = (elementName:string, options:CustomDataElementOptions={}) =>
-    (ctor: CustomElementConstructor) => {    
-    options.stateIdProperty && setProp(ctor, "stateIdProperty", options.stateIdProperty);
-    options.eventsListenAt && setProp(ctor, "eventsListenAt", options.eventsListenAt);
-    customDataElements.define(elementName, ctor);
-};
-
-
-
-interface DataPropertyOptions {
-    changeEvent:string
-}
-/**
- * A property decorator that tags a class property
- * as a state property.
- * 
- * Options allow for setting the change event name.
- * @param options 
- */
-const dataProperty = (options?:DataPropertyOptions):any =>
-    (prototype: any, propertyName: string) => {
-        if (prototype.constructor.dataProperties === DataElement.dataProperties) {
-            prototype.constructor.dataProperties = {};
-        }
-        (prototype.constructor as DataElementCtor).dataProperties[propertyName] = {
-            changeEvent: options ? options.changeEvent : `${propertyName}-changed`
-        };
-    };
 
 
 const elementConnected = (el:DataElement) => {
