+++ packages/DataElement/src/index.ts
@@ -1,16 +1,7 @@
 export {
     customDataElements,
-    customDataElement,
-    DataElement,
-    dataProperty,
-    event
+    DataElement
 } from "./DataElement";
 export { 
     StateChange
 } from "@domx/statechange";
-export {
-    applyEventMapLogging
-} from "@domx/eventmap/applyEventMapLogging";
-export {
-    applyImmerToStateChange
-} from "@domx/statechange/applyImmerToStateChange";
