+++ packages/DataElement/src/middleware.ts
@@ -1,0 +1,3 @@
+export { applyEventMapLogging } from "@domx/eventmap/applyEventMapLogging";
+export { applyImmerToStateChange } from "@domx/statechange/applyImmerToStateChange";
+export { applyDataElementRdtLogging } from "./applyDataElementRdtLogging";
