+++ packages/DataElement/tsconfig.json
@@ -12,5 +12,10 @@
   ],
   "exclude": [
     "./src/**/*.spec.ts"
+  ],
+  "references": [
+    { "path": "../EventMap"},
+    { "path": "../StateChange" },
+    { "path": "../Middleware" }
   ]
 }
